+++ extension/cart/getCheckoutUrlFromMagento.js
@@ -4,7 +4,7 @@ const moment = require('moment')
 const MagentoError = require('../models/Errors/MagentoEndpointError')
 const ResponseParser = require('../helpers/MagentoResponseParser')
 const InvalidCallError = require('../models/Errors/InvalidCallError')
-const util = require('util')
+const { warn: logMageWarn, debug: logMageDebug } = require('../models/Logs/mage')
 
 /**
  * @typedef {Object} getCheckoutUrlFromMagentoInput
@@ -22,17 +22,16 @@ const util = require('util')
 module.exports = function (context, input, cb) {
   const request = context.tracedRequest('magento-cart-extension:getCheckoutUrlFromMagento', { log: true })
   const cartUrl = context.config.magentoUrl + '/carts'
-  const log = context.log
   const allowSelfSignedCertificate = context.config.allowSelfSignedCertificate
   const accessToken = input.token
   const cartId = input.cartId
 
   if (!cartId) {
-    log.error('Output key "cartId" is missing')
+    context.log.error('Output key "cartId" is missing')
     return cb(new InvalidCallError())
   }
 
-  getCheckoutUrlFromMagento(request, accessToken, cartId, cartUrl, log, !allowSelfSignedCertificate, (err, result) => {
+  getCheckoutUrlFromMagento(request, accessToken, cartId, cartUrl, context, !allowSelfSignedCertificate, (err, result) => {
     if (err) return cb(err)
 
     // Add additional query parameters for Google Analytics in Webcheckout
@@ -56,11 +55,11 @@ module.exports = function (context, input, cb) {
  * @param {string} accessToken
  * @param {number|string} cartId
  * @param {string} cartUrl
- * @param {Logger} log
+ * @param {context} context
  * @param {boolean} rejectUnauthorized
  * @param {StepCallback} cb
  */
-function getCheckoutUrlFromMagento (request, accessToken, cartId, cartUrl, log, rejectUnauthorized, cb) {
+function getCheckoutUrlFromMagento (request, accessToken, cartId, cartUrl, context, rejectUnauthorized, cb) {
   const options = {
     baseUrl: cartUrl,
     uri: cartId + '/checkoutUrl',
@@ -74,25 +73,16 @@ function getCheckoutUrlFromMagento (request, accessToken, cartId, cartUrl, log, 
     if (err) return cb(err)
 
     if (!res.body) {
-      log.error(options, `Got empty body from magento. Request result: ${res}`)
+      context.log.error(options, `Got empty body from magento. Request result: ${res}`)
       return cb(new MagentoError())
     }
 
     if (res.statusCode !== 200 || !res.body.url) {
-      log.error(`Got ${res.statusCode} from magento: ${ResponseParser.extractMagentoError(res.body)}`)
+      logMageWarn(context, res, ResponseParser.extractMagentoError(res.body))
       return cb(new MagentoError())
     }
 
-    log.debug(
-      {
-        duration: new Date() - requestStart,
-        statusCode: res.statusCode,
-        request: util.inspect(options, true, 5),
-        response: util.inspect(res.body, true, 5)
-      },
-      'Request to Magento: getCheckoutUrlFromMagento'
-    )
-
+    logMageDebug(context, requestStart, options, res, 'Request to Magento: getCheckoutUrlFromMagento')
     cb(null, res.body)
   })
 }