+++ extension/cart/setCartCustomer_v1.js
@@ -1,6 +1,6 @@
 const MagentoError = require('../models/Errors/MagentoEndpointError')
 const ResponseParser = require('../helpers/MagentoResponseParser')
-const util = require('util')
+const { error: logMageError, debug: logMageDebug } = require('../models/Logs/mage')
 
 /**
  * @typedef {Object} SetCartCustomer_v1Input
@@ -16,17 +16,16 @@ const util = require('util')
 module.exports = function (context, input, cb) {
   const cartUrl = context.config.magentoUrl + '/carts'
   const cartId = input.cartId
-  const log = context.log
   const allowSelfSignedCertificate = context.config.allowSelfSignedCertificate
 
   // cart must exist to be able to assign a customer for it
   if (!cartId) {
     // this can happen pretty regularly, as guest carts are only created if products are added
-    log.debug('setCartCustomer is skipped, because no valid cartId was given')
+    context.log.debug('setCartCustomer is skipped, because no valid cartId was given')
     return cb()
   }
 
-  assignCartCustomer(context.tracedRequest('magento-cart-extension:setCartCustomer', { log: true }), input.token, cartId, cartUrl, log, !allowSelfSignedCertificate, (err) => {
+  assignCartCustomer(context.tracedRequest('magento-cart-extension:setCartCustomer', { log: true }), input.token, cartId, cartUrl, context, !allowSelfSignedCertificate, (err) => {
     if (err) return cb(err)
     cb(null, { messages: null })
   })
@@ -37,11 +36,11 @@ module.exports = function (context, input, cb) {
  * @param {string} accessToken
  * @param {(number|string)} cartId - can be cart ID for guest or "me" for customer
  * @param {string} cartUrl
- * @param {Logger} log
+ * @param {context} context
  * @param {boolean} rejectUnauthorized
  * @param {StepCallback} cb
  */
-function assignCartCustomer (request, accessToken, cartId, cartUrl, log, rejectUnauthorized, cb) {
+function assignCartCustomer (request, accessToken, cartId, cartUrl, context, rejectUnauthorized, cb) {
   const options = {
     baseUrl: cartUrl,
     uri: cartId.toString() + '/customer',
@@ -54,19 +53,11 @@ function assignCartCustomer (request, accessToken, cartId, cartUrl, log, rejectU
   request.post(options, (err, res) => {
     if (err) return cb(err)
     if (res.statusCode !== 200) {
-      log.error(`Got ${res.statusCode} from magento: ${ResponseParser.extractMagentoError(res.body)}`)
+      logMageError(context, res, ResponseParser.extractMagentoError(res.body))
       return cb(new MagentoError())
     }
 
-    log.debug(
-      {
-        duration: new Date() - requestStart,
-        statusCode: res.statusCode,
-        request: util.inspect(options, true, 5),
-        response: util.inspect(res.body, true, 5)
-      },
-      'Request to Magento: setCartCustomer'
-    )
+    logMageDebug(context, requestStart, options, res, 'Request to Magento: setCartCustomer')
     cb()
   })
 }