+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- log levels for addCartItems and getCheckoutUrlFromMagento to be warnings instead of errors
+
 ## [1.2.1] - 2019-04-12
 ### Fixed
 - issue when a cart was deleted from Magento database, now it creates a new one instead of throwing a hard error